@@ -9,29 +9,30 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~11.0.5",
-    "@angular/common": "~11.0.5",
-    "@angular/compiler": "~11.0.5",
-    "@angular/core": "~11.0.5",
-    "@angular/forms": "~11.0.5",
-    "@angular/platform-browser": "~11.0.5",
-    "@angular/platform-browser-dynamic": "~11.0.5",
-    "@angular/router": "~11.0.5",
-    "bootstrap": "5.0.0-beta1",
-    "bootstrap-icons": "1.2.2",
-    "ngx-highlightjs": "^4.1.3",
-    "rxjs": "~6.6.0",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.10.2"
+    "@angular/animations": "~16.2.6",
+    "@angular/common": "~16.2.6",
+    "@angular/compiler": "~16.2.6",
+    "@angular/core": "~16.2.6",
+    "@angular/forms": "~16.2.6",
+    "@angular/platform-browser": "~16.2.6",
+    "@angular/platform-browser-dynamic": "~16.2.6",
+    "@angular/router": "~16.2.6",
+    "bootstrap": "~5.3.2",
+    "bootstrap-icons": "~1.11.1",
+    "ci": "^2.3.0",
+    "ngx-highlightjs": "~10.0.0",
+    "rxjs": "~7.8.1",
+    "tslib": "~2.6.2",
+    "zone.js": "~0.13.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1100.5",
-    "@angular/cli": "~11.0.5",
-    "@angular/compiler-cli": "~11.0.5",
-    "@types/jasmine": "~3.6.0",
-    "codelyzer": "^6.0.0",
-    "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "~4.0.2"
+    "@angular-devkit/build-angular": "~16.2.3",
+    "@angular/cli": "~16.2.3",
+    "@angular/compiler-cli": "~16.2.6",
+    "@types/jasmine": "~4.3.6",
+    "codelyzer": "~6.0.2",
+    "ts-node": "~10.9.1",
+    "tslint": "~6.1.3",
+    "typescript": "~4.9.3"
   }
 }
