@@ -1,24 +1,23 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { HttpClientModule } from '@angular/common/http';
-import { HighlightModule, HIGHLIGHT_OPTIONS } from 'ngx-highlightjs';
+import { HIGHLIGHT_OPTIONS } from 'ngx-highlightjs';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { HealthComponent } from './health/health.component';
-import { DocsComponent } from './docs/docs.component';
 import { SubscriptionsComponent } from './subscriptions/subscriptions.component';
 import { HomeComponent } from './home/home.component';
 
 @NgModule({
-  declarations: [AppComponent, HealthComponent, DocsComponent, SubscriptionsComponent, HomeComponent],
-  imports: [BrowserModule, AppRoutingModule, HttpClientModule, HighlightModule],
+  declarations: [AppComponent, HealthComponent, SubscriptionsComponent, HomeComponent],
+  imports: [BrowserModule, AppRoutingModule, HttpClientModule],
   providers: [
     {
       provide: HIGHLIGHT_OPTIONS,
       useValue: {
         coreLibraryLoader: () => import('highlight.js/lib/core'),
-        lineNumbersLoader: () => import('highlightjs-line-numbers.js'),
+        lineNumbersLoader: () => import('ngx-highlightjs/line-numbers'),
         languages: { csharp: () => import('highlight.js/lib/languages/csharp') }
       }
     }],
